Simplified vue
I'd like to get this down to a single page if possible to really see how
it works, it's a bit obscured when initialised. Multiple files and it's
not clear what's totally necessary.

Did some pruning and have a basic set of (so far) minimal files to get a
some words to show in a browser.